+++ tests/file_test.rs
@@ -1,0 +1,51 @@
+use clipboard_rs::{Clipboard, ClipboardContext, ContentFormat};
+
+#[cfg(target_os = "macos")]
+const TMP_PATH: &str = "/tmp/";
+#[cfg(target_os = "windows")]
+const TMP_PATH: &str = "C:\\Windows\\Temp\\";
+#[cfg(all(
+    unix,
+    not(any(
+        target_os = "macos",
+        target_os = "ios",
+        target_os = "android",
+        target_os = "emscripten"
+    ))
+))]
+const TMP_PATH: &str = "/tmp/";
+
+#[test]
+fn test_file() {
+    let ctx = ClipboardContext::new().unwrap();
+
+    let files = get_files();
+
+    ctx.set_files(files).unwrap();
+
+    let types = ctx.available_formats().unwrap();
+    println!("{:?}", types);
+
+    let has = ctx.has(ContentFormat::Files);
+    assert_eq!(has, true);
+
+    let files = ctx.get_files().unwrap();
+    assert_eq!(files.len(), 2);
+
+    for file in files {
+        println!("{:?}", file);
+    }
+
+    ctx.clear().unwrap();
+
+    let has = ctx.has(ContentFormat::Files);
+    assert_eq!(has, false);
+}
+
+fn get_files() -> Vec<String> {
+    let test_file1 = format!("{}clipboard_rs_test_file1.txt", TMP_PATH);
+    let test_file2 = format!("{}clipboard_rs_test_file2.txt", TMP_PATH);
+    std::fs::write(&test_file1, "hello world").unwrap();
+    std::fs::write(&test_file2, "hello world").unwrap();
+    vec![test_file1, test_file2]
+}