+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # Changelog
+
+## v0.1.1 (2024-03-04) [unreleased]
+
+- Feature: Add a option to `getFiles` or `setFiles`
+- Feature: Add a option to `get` or `set` multi items
+- Fix: make `WatcherShutdown` public